// frontend/generated/models/FileDto.ts
export interface FileDto {
    /**
     * 
     * @type {number}
     * @memberof FileDto
     */
    fileId?: number;
    /**
     * 
     * @type {AccountDto}
     * @memberof FileDto
     */
    uploader?: AccountDto;
    /**
     * 
     * @type {string}
     * @memberof FileDto
     */
    fileName?: string;
    /**
     * 
     * @type {string}
     * @memberof FileDto
     */
    fileUrl?: string;
    /**
     * 
     * @type {string}
     * @memberof FileDto
     */
    fileType?: string;
    /**
     * 
     * @type {number}
     * @memberof FileDto
     */
    size?: number;
    /**
     * 
     * @type {boolean}
     * @memberof FileDto
     */
    isVisible?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof FileDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {number}
     * @memberof FileDto
     */
    messageId?: number;
    /**
     * 
     * @type {number}
     * @memberof FileDto
     */
    proposalId?: number;
    /**
     * 
     * @type {number}
     * @memberof FileDto
     */
    projectId?: number;
    /**
     * 
     * @type {number}
     * @memberof FileDto
     */
    milestoneId?: number;
}