// frontend/generated/models/MilestoneDto.ts
export interface MilestoneDto {
    /**
     * 
     * @type {number}
     * @memberof MilestoneDto
     */
    milestoneId?: number;
    /**
     * 
     * @type {number}
     * @memberof MilestoneDto
     */
    projectId?: number;
    /**
     * 
     * @type {string}
     * @memberof MilestoneDto
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof MilestoneDto
     */
    title?: string;
    /**
     * 
     * @type {number}
     * @memberof MilestoneDto
     */
    budgetRatio?: number;
    /**
     * 
     * @type {number}
     * @memberof MilestoneDto
     */
    contractualBudget?: number;
    /**
     * 
     * @type {Date}
     * @memberof MilestoneDto
     */
    deadline?: Date;
    /**
     * 
     * @type {MilestoneStatusDto}
     * @memberof MilestoneDto
     */
    status?: MilestoneStatusDto;
    /**
     * 
     * @type {MilestoneFundStatusDto}
     * @memberof MilestoneDto
     */
    fundStatus?: MilestoneFundStatusDto;
    /**
     * 
     * @type {boolean}
     * @memberof MilestoneDto
     */
    isVisible?: boolean;
    /**
     * 
     * @type {Array<FileDto>}
     * @memberof MilestoneDto
     */
    deliverables?: Array<FileDto>;
    /**
     * 
     * @type {Date}
     * @memberof MilestoneDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof MilestoneDto
     */
    updatedAt?: Date;
}